LPS-144606 Move BNDSettings.java to util package
