+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/BaseSourceCheck.java
@@ -27,7 +27,6 @@ import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.tools.ToolsUtil;
-import com.liferay.source.formatter.BNDSettings;
 import com.liferay.source.formatter.SourceFormatterExcludes;
 import com.liferay.source.formatter.SourceFormatterMessage;
 import com.liferay.source.formatter.checks.util.JSPSourceUtil;
@@ -35,6 +34,7 @@ import com.liferay.source.formatter.checks.util.SourceUtil;
 import com.liferay.source.formatter.processors.JSPSourceProcessor;
 import com.liferay.source.formatter.processors.JavaSourceProcessor;
 import com.liferay.source.formatter.processors.SourceProcessor;
+import com.liferay.source.formatter.util.BNDSettings;
 import com.liferay.source.formatter.util.CheckType;
 import com.liferay.source.formatter.util.FileUtil;
 import com.liferay.source.formatter.util.SourceFormatterCheckUtil;
