+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/JavaModuleExposureCheck.java
@@ -14,12 +14,12 @@
 
 package com.liferay.source.formatter.checks;
 
-import com.liferay.source.formatter.BNDSettings;
 import com.liferay.source.formatter.checks.util.BNDSourceUtil;
 import com.liferay.source.formatter.parser.JavaClass;
 import com.liferay.source.formatter.parser.JavaParameter;
 import com.liferay.source.formatter.parser.JavaSignature;
 import com.liferay.source.formatter.parser.JavaTerm;
+import com.liferay.source.formatter.util.BNDSettings;
 
 import java.io.IOException;
