+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/JavaOSGiReferenceCheck.java
@@ -19,12 +19,12 @@ import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.source.formatter.BNDSettings;
 import com.liferay.source.formatter.checks.util.JavaSourceUtil;
 import com.liferay.source.formatter.parser.JavaClass;
 import com.liferay.source.formatter.parser.JavaClassParser;
 import com.liferay.source.formatter.parser.JavaTerm;
 import com.liferay.source.formatter.parser.ParseException;
+import com.liferay.source.formatter.util.BNDSettings;
 import com.liferay.source.formatter.util.FileUtil;
 
 import java.io.File;
