+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/JavaProviderTypeAnnotationCheck.java
@@ -16,10 +16,10 @@ package com.liferay.source.formatter.checks;
 
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.source.formatter.BNDSettings;
 import com.liferay.source.formatter.checks.util.BNDSourceUtil;
 import com.liferay.source.formatter.parser.JavaClass;
 import com.liferay.source.formatter.parser.JavaTerm;
+import com.liferay.source.formatter.util.BNDSettings;
 
 import java.io.IOException;
