+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/PackageinfoBNDExportPackageCheck.java
@@ -17,7 +17,7 @@ package com.liferay.source.formatter.checks;
 import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.source.formatter.BNDSettings;
+import com.liferay.source.formatter.util.BNDSettings;
 
 import java.io.IOException;
